requirements.txt
requests
sklearn
plotly.express